+++ packages/typedoc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/typedoc
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.0.1
+
 ## 1.0.0
 
 ### Major Changes