+++ packages/vitepress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/vitepress
 
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.0.1
+
 ## 0.2.17
 
 ### Patch Changes